Fixed bug in TriOsc: width needed to be a float.

+++ PqExtra.h
@@ -110,7 +110,7 @@ public:
   void _update(float t);
 
   unsigned long _period;
-  unsigned long _width;
+  float _width;
   unsigned long _startTime;
   float _value;
 };
